# api/public.openapi.yaml
openapi: 3.0.3
info:
  title: Pratu Public API
  version: 0.2.0
  description: |
    The tenant-facing API. Every request is addressed to a tenant hostname
    (`{slug}.{base_domain}` or a custom domain): the Host header selects the
    tenant, and each tenant is its own OIDC issuer.

    **Flows** come in two variants. *API flows* (`/api` creation endpoints)
    are for native apps: they return `session_token` values and need no
    CSRF. *Browser flows* (`/browser` creation endpoints) set an HTTP-only
    CSRF cookie, return a `csrf_token` bound to the flow, require it on
    every submission, and set the `pratu_session` cookie instead of
    returning a session token.

    **Browser flows are content-negotiated.** A client that prefers
    `text/html` — or posts `application/x-www-form-urlencoded` — is driven
    by 303 redirects to the tenant's configured screens: flow creation
    lands on the screen with `?flow=`, a failed submission returns there
    with its messages persisted on the flow, a failure with no flow left
    to return to lands on the error screen with `?code=`, and a completed
    flow lands on `return_to` (validated against the tenant's allow-list)
    or the tenant's default return URL. A client that asks for
    `application/json` gets the JSON responses documented below,
    unchanged. Screens re-read the flow with
    `GET /self-service/flows/{id}`.

    Cookie-authenticated requests to state-changing session endpoints
...

...
  license:
...
servers:
...
paths:
...
  /.well-known/jwks.json:
    get:
      tags: [oauth2]
      summary: "This tenant's signing keys (active and retired)"
      responses:
        "200": { description: JSON Web Key Set, content: { application/json: { schema: { type: object } } } }
        "503": { $ref: "#/components/responses/Error" }
  /oauth2/auth:
    get:
      tags: [oauth2]
      summary: Authorization endpoint (RFC 6749)
      description: |
        Validates the request and 302-redirects the browser to the tenant's
        configured `login_url` with a `login_challenge` parameter.
      responses:
        "302": { description: Redirect to the tenant login UI }
        "400": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
  /oauth2/token:
    post:
      tags: [oauth2]
      summary: Token endpoint (authorization_code + PKCE, refresh_token)
      description: |
        Client auth: `client_secret_basic`, `client_secret_post`, or `none`
        with PKCE for public clients. Access tokens are 15-minute RS256
        JWTs; refresh tokens rotate on every use and reuse revokes the
        whole chain.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema: { type: object }
      responses:
        "200": { description: Token response (RFC 6749 §5.1), content: { application/json: { schema: { type: object } } } }
        "400": { description: OAuth2 error, content: { application/json: { schema: { type: object } } } }
        "401": { description: Client authentication failed }
  /oauth2/introspect:
    post:
      tags: [oauth2]
      summary: Token introspection (RFC 7662; client-authenticated)
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema: { type: object }
      responses:
        "200": { description: Introspection response, content: { application/json: { schema: { type: object } } } }
        "401": { description: Client authentication failed }
  /oauth2/revoke:
    post:
      tags: [oauth2]
      summary: Token revocation (RFC 7009; client-authenticated)
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema: { type: object }
      responses:
        "200": { description: Revoked (or token unknown; per RFC both are 200) }
        "401": { description: Client authentication failed }
components:
  securitySchemes:
    sessionToken:
      type: apiKey
      in: header
      name: X-Session-Token
      description: Opaque session token (also accepted as `Authorization Bearer`)
    sessionCookie:
      type: apiKey
      in: cookie
      name: pratu_session
      description: Browser-flow session cookie; state-changing calls also need `X-CSRF-Token`
...
  responses:
...
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/APIError" }
    RateLimited:
      description: Rate limited; Retry-After indicates when the window rolls over
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/APIError" }
  schemas:
    APIError:
      type: object
      properties:
        error:
          type: object
          properties:
            message: { type: string }
            details:
              type: array
              items: { type: string }
...